+++ CHANGELOG.md
@@ -46,6 +46,11 @@ New devices:
 - CR Smart Home Smoke sensor
 - CR Smart Home Socket
 - CR Smart Home Water leak detector
+- IKEA TRADFRI LED bulb E14 470 lumen, dimmable, white spectrum, opal white
+- Zemismart ZM-L03E-Z 3 gang with neutral wire switch
+- Zemismart RGB LED downlight
+- Zemismart Curtain/roller blind switch
+- Bitron Wireless wall thermostat with relay
 
 ## 0.0.20
 - Properly handle heartbeat message from IKEA TRADFRI Remote Control